requirements.txt
streamlit
tensorflow-cpu==2.20.0
numpy
Pillow
gdown
